hsql scripotd debugged
git-svn-id: http://datacenter.gradsoft.ua/repos/Jungle/jungle.public@14883 e6263924-d73c-0410-abf9-b8ad5fbdd136

+++ facilities/configuration/sql/hsql/drop_configuration_hsql.sql
@@ -1,3 +1,3 @@
 
-drop table jungle_configitems;
+drop table jungle_configitems if exists;
